requirements.txt
pandas
scikit-learn
nltk
groq
python-dotenv
